fix: restustructure routes
